支持内容删除 自定义

+++ main.js
@@ -29,17 +29,20 @@ cmd.command('init').description('初始化模板').action(async (args) => {
     await downGit(chooseMap[chooses.type].url, args)
     log.success('项目创建成功\n')
     try {
-        await installDependencies(path.join(cwd, args), 'npm')
         // 生成对应的manifest文件
-        await generateManifest(chooses, path.join(__dirname, `${args}/src/manifest.json`))
+        await generateManifest(chooses, path.join(cwd, `${args}/src/manifest.json`))
         // 删除对应不需要的目录
         for (let index = 0; index < defaultKeys.length; index++) {
             const name = defaultKeys[index];
             let { dir } = defaultKeysMap[name];
-            let needRmPath = path.join(__dirname, `${args}/src/${dir || name}`)
-            // console.log('开始删除', needRmPath);
-            await nodeUtil.doShellCmd(`rm -rf ${needRmPath}`)
+            if (!chooses[name]) {
+                let needRmPath = path.join(cwd, `${args}/src/${dir || name}`)
+                // console.log('开始删除', needRmPath);
+                await nodeUtil.doShellCmd(`rm -rf ${needRmPath}`)
+            }
         }
+        // 自动安装依赖
+        await installDependencies(path.join(cwd, args), 'npm')
         log.success('依赖安装成功 请运行 npm run build 生成dist目录，然后将此目录安装进扩展目录中\n')
         log.success('安装扩展方法可见 https://juejin.cn/post/7148618712646418439\n')
         process.exit(0)
